requirements.txt
python-telegram-bot[all]==20.1
openai==1.3.4
pytest==6.1.2